Made SpawnLocation and Seats appear as normally created.
So normally in studio when you create a spawnlocation it has a square shape and has a the spawn decal on top of it.
Also Seats and VehicleSeats have their brickcolor set to black when created in studio.

Also for me at least in F3X BTools when spawning these parts you easily confuse them with regular parts.